refactor: перенесла файл App в папку app
