+++ packages/client/src/app/App.tsx
@@ -1,5 +1,6 @@
 import { useEffect } from 'react'
 import './App.css'
+import { AppRouter } from '@/app/providers/AppRouter'
 
 function App() {
   useEffect(() => {
@@ -12,7 +13,11 @@ function App() {
 
     fetchServerData()
   }, [])
-  return <div className="App">Вот тут будет жить ваше приложение :)</div>
+  return (
+    <div className="App">
+      <AppRouter />
+    </div>
+  )
 }
 
 export default App
